basic functionality done
