+++ src/components/Board/Board.constants.js
@@ -1,0 +1,14 @@
+import asteroid1 from '../../assets/images/asteroid1.png'
+import asteroid2 from '../../assets/images/asteroid2.png'
+
+export const ASTEROID_GENERATE_RATE = 1000;
+export const UPDATE_RATE = 200;
+export const START_POINT = 150;
+export const PLAYER_MOVEMENT_AMOUNT = 15;
+export const CHECK_KEYS_INTERVAL = 50;
+export const KEY_INACTIVITY_TIMEOUT = 3000;
+
+export const ASTEROID_IMAGES = [
+    asteroid1, asteroid2
+  ]
+
